requirements.txt
 pandas==2.2.3
 plotly_express==6.2.0
 streamlit==1.46.1